Add support for loading plain HTML and QML.
Plain means the content does not contain a webvfx.renderReady(true)
call. The new behavior requires that the filename ends with .plain.html
or .plain.qml. There is a WEBVFX_FILENAMEEXT define that inverses the
behavior such that .html and .qml load without webvfx.renderReady(true)
and a .webvfx.html or .webvfx.qml is required to make rendering wait for
renderReady(true).