A better way to setup the consumer-stopping listener.
This is to prevent a deadlock on the rendering thread waiting for Qt to
render while the main Qt thread has called mlt_consumer_stop() and
waiting for the rendering thread to stop. Previously, the "fix" required
the application to remember to set a 'consumer' MLT data property on the
webvfx service, but that is not reliable and did not work at all with
the MLT XML producer. Now, the mlt_consumer base class sets a consumer
property on the frames it sends to webvfx.